+++ packages/components/src/components/button-pure/button-pure-styles.spec.ts
@@ -1,5 +1,6 @@
 import { getComponentCss } from './button-pure-styles';
 import type { AlignLabel, BreakpointCustomizable, LinkButtonIconName, TextSize, Theme } from '../../types';
+import { validateCssAndMatchSnapshot } from '../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
   const breakpointCustomizableBoolean = { base: true, xs: false, s: true, m: false, l: true, xl: false };
@@ -239,7 +240,7 @@ describe('getComponentCss()', () => {
   ])(
     'should return correct css for %j',
     ({ icon, iconSource, active, isLoading, isDisabledOrLoading, stretch, size, hideLabel, alignLabel, theme }) => {
-      expect(
+      validateCssAndMatchSnapshot(
         getComponentCss(
           icon,
           iconSource,
@@ -252,7 +253,7 @@ describe('getComponentCss()', () => {
           alignLabel,
           theme
         )
-      ).toMatchSnapshot();
+      );
     }
   );
 });