+++ packages/components/src/components/common/state-message/state-message-styles.spec.ts
@@ -1,5 +1,6 @@
 import { getFunctionalComponentStateMessageStyles } from './state-message-styles';
 import { getCss } from '../../../utils';
+import { validateCssAndMatchSnapshot } from '../../../../tests/unit/helpers';
 
 describe('getFunctionalComponentStateMessageStyles()', () => {
   it.each<Parameters<typeof getFunctionalComponentStateMessageStyles>>([
@@ -10,6 +11,6 @@ describe('getFunctionalComponentStateMessageStyles()', () => {
     ['dark', 'success'],
     ['dark', 'error'],
   ])('should return correct JssStyle for theme: %s and state: %s', (...args) => {
-    expect(getCss(getFunctionalComponentStateMessageStyles(...args))).toMatchSnapshot();
+    validateCssAndMatchSnapshot(getCss(getFunctionalComponentStateMessageStyles(...args)));
   });
 });