+++ packages/components/src/components/content-wrapper/content-wrapper-styles.spec.ts
@@ -1,10 +1,11 @@
 import { getComponentCss } from './content-wrapper-styles';
+import { validateCssAndMatchSnapshot } from '../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([['full'], ['fluid'], ['extended'], ['basic'], ['narrow']])(
     'should return correct css for width: %s',
     (...args) => {
-      expect(getComponentCss(...args)).toMatchSnapshot();
+      validateCssAndMatchSnapshot(getComponentCss(...args));
     }
   );
 });