+++ packages/components/src/components/multi-select/multi-select-option/multi-select-option-styles.spec.ts
@@ -1,10 +1,11 @@
 import { getComponentCss } from './multi-select-option-styles';
+import { validateCssAndMatchSnapshot } from '../../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([['light'], ['dark']])(
     'should return correct css for theme: %s',
     (...args) => {
-      expect(getComponentCss(...args)).toMatchSnapshot();
+      validateCssAndMatchSnapshot(getComponentCss(...args));
     }
   );
 });