+++ packages/components/src/components/pin-code/pin-code-styles.spec.ts
@@ -1,4 +1,5 @@
 import { getComponentCss } from './pin-code-styles';
+import { validateCssAndMatchSnapshot } from '../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
@@ -18,7 +19,7 @@ describe('getComponentCss()', () => {
   ])(
     'should return correct css for hideLabel: %o, state: %s, isDisabled: %s, isLoading: %s, isWithinForm: %s, length: %s and theme: %s',
     (...args) => {
-      expect(getComponentCss(...args)).toMatchSnapshot();
+      validateCssAndMatchSnapshot(getComponentCss(...args));
     }
   );
 });