+++ packages/components/src/components/text-list/text-list-item/text-list-item-styles.spec.ts
@@ -1,7 +1,8 @@
 import { getComponentCss } from './text-list-item-styles';
+import { validateCssAndMatchSnapshot } from '../../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([[]])('should return correct css', (...args) => {
-    expect(getComponentCss(...args)).toMatchSnapshot();
+    validateCssAndMatchSnapshot(getComponentCss(...args));
   });
 });