+++ packages/components/src/components/text-list/text-list/text-list-styles.spec.ts
@@ -1,4 +1,5 @@
 import { getComponentCss } from './text-list-styles';
+import { validateCssAndMatchSnapshot } from '../../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
@@ -9,6 +10,6 @@ describe('getComponentCss()', () => {
     ['numbered', 'dark'],
     ['alphabetically', 'dark'],
   ])('should return correct css for type: %s and theme: %s', (...args) => {
-    expect(getComponentCss(...args)).toMatchSnapshot();
+    validateCssAndMatchSnapshot(getComponentCss(...args));
   });
 });