+++ packages/components/src/components/text/text-styles.spec.ts
@@ -1,4 +1,5 @@
 import { getComponentCss } from './text-styles';
+import { validateCssAndMatchSnapshot } from '../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
@@ -18,7 +19,7 @@ describe('getComponentCss()', () => {
   ])(
     'should return correct css for size: %j, weight: %s, align: %s, color: %s, ellipsis: %o and theme: %s',
     (...args) => {
-      expect(getComponentCss(...args)).toMatchSnapshot();
+      validateCssAndMatchSnapshot(getComponentCss(...args));
     }
   );
 });